Split up raft binary tests
