+++ vault/external_tests/raft/raft_test.go
@@ -442,76 +442,7 @@ func TestRaft_Configuration(t *testing.T) {
 	t.Parallel()
 	cluster, _ := raftCluster(t, nil)
 	defer cluster.Cleanup()
-	testRaft_Configuration(t, cluster)
-}
-
-// TestRaft_Configuration_Docker is a variant of TestRaft_Configuration that
-// uses docker containers for the vault nodes.
-func TestRaft_Configuration_Docker(t *testing.T) {
-	t.Parallel()
-	binary := os.Getenv("VAULT_BINARY")
-	if binary == "" {
-		t.Skip("only running docker test when $VAULT_BINARY present")
-	}
-	opts := &docker.DockerClusterOptions{
-		ImageRepo: "hashicorp/vault",
-		// We're replacing the binary anyway, so we're not too particular about
-		// the docker image version tag.
-		ImageTag:    "latest",
-		VaultBinary: binary,
-		ClusterOptions: testcluster.ClusterOptions{
-			VaultNodeConfig: &testcluster.VaultNodeConfig{
-				LogLevel: "TRACE",
-				// If you want the test to run faster locally, you could
-				// uncomment this performance_multiplier change.
-				//StorageOptions: map[string]string{
-				//	"performance_multiplier": "1",
-				//},
-			},
-		},
-	}
-	cluster := docker.NewTestDockerCluster(t, opts)
-	defer cluster.Cleanup()
-	testRaft_Configuration(t, cluster)
-
-	if err := cluster.AddNode(context.TODO(), opts); err != nil {
-		t.Fatal(err)
-	}
-	testRaft_Configuration(t, cluster)
-}
-
-func testRaft_Configuration(t *testing.T, cluster testcluster.VaultCluster) {
-	client := cluster.Nodes()[0].APIClient()
-	secret, err := client.Logical().Read("sys/storage/raft/configuration")
-	if err != nil {
-		t.Fatal(err)
-	}
-	servers := secret.Data["config"].(map[string]interface{})["servers"].([]interface{})
-	found := make(map[string]struct{})
-	for _, s := range servers {
-		server := s.(map[string]interface{})
-		nodeID := server["node_id"].(string)
-		leader := server["leader"].(bool)
-		switch nodeID {
-		case "core-0":
-			if !leader {
-				t.Fatalf("expected server to be leader: %#v", server)
-			}
-		default:
-			if leader {
-				t.Fatalf("expected server to not be leader: %#v", server)
-			}
-		}
-
-		found[nodeID] = struct{}{}
-	}
-	expected := map[string]struct{}{}
-	for i := range cluster.Nodes() {
-		expected[fmt.Sprintf("core-%d", i)] = struct{}{}
-	}
-	if diff := deep.Equal(expected, found); len(diff) > 0 {
-		t.Fatalf("configuration mismatch, diff: %v", diff)
-	}
+	Raft_Configuration_Test(t, cluster)
 }
 
 func TestRaft_ShamirUnseal(t *testing.T) {
